reorganized index appearing
