+++ Src/style.css
@@ -1,6 +1,5 @@
 body{
-    background-image: url('astrolove_logo.jpg');
-   
+    background-image: url("Design/astrolove_logo.jpg");
     background-repeat: no-repeat;
     background-position: center;
     background-color: #220c34;
